Consolidate VirtualTable into DataTable
Single DataTable component with built-in virtual scrolling for all
table views (Explorer, Savings, TableWidget). Removes the separate
VirtualTable — same behaviour everywhere.

Falls back to rendering all rows when the virtualizer can't measure
the container (jsdom in tests).
